Initialize Insn _block data member
This isn't really necessary, as it's always initialized in the
constructor, but for the purpose of avoiding uninitialized data, it's
easier to just make sure all data members have an explicitly
initialized declaration.  Constructor initialization will override the
initialization in the declaration, so doing this will never result in
inefficiency.