Move argument / result vectors into Insn class
Instead of having various subclasses contain their own argument and/or
result register vectors, and using virtual methods to control access
to them, simply put argument / result vector into the base Insn class.

For classes that don't use arguments and/or results, these will simply
be empty vectors.

This allows a lot of code simplification and refactoring, and in
practice is probably not really much of a space waste, as the majority
of instruction will probably be instances subclasses with such vectors
anyway.