Rework handling of dominator info in basic blocks
Instead of linking between internal DomTreeNode nodes in basic blocks,
link to the basic blocks themselves.  The actual pointers are still in
DomTreeNode members, so that we can share code between forward and
backward dominator trees.

This means that internal dominator-related methods are moved from
the BB::DomTreeNode class directly into the BB class.

Doing things this way is slightly more verbose and requires a bit more
explicit usage of pointers-to-members, but it's overall cleaner and
avoids the use of slightly dodgy pointer arithmetic.

It also means that we have a direct list of BB* dominatees available,
which will be useful in the future.