Make SrcFileInput::skip work for non-identifier "keywords"
That is, the "keyword" argument doesn't have to begin with a valid
identifier character.  The only difference in the way non-keywords are
handled is that they can be immediately (without any whitespace)
followed by an identifier.

This change also affects any other interfacts like
SrcFileInput::expect which call SrcFileInput::skip.